Use static service name and populate k8s resource as additional attribute (#2)
## Which problem is this PR solving?

When running kspan it generates high cardinality values for `service.name` attribute. This makes it hard to work with honeycomb as we create a dataset for each of these. This PR switches around the `service.name` attribute and populates the static value `kspan` which should then create a single dataset in honeycomb. It then populates what used to be in the `service.name` attribute into `k8s.service`

## Short description of the changes

Retain all of the existing flow and just massage the span before emitting it

## How to verify that this has the expected result

Unit tests